feat!: global bins should be created only in predefined locations
