+++ packages/config/package.json
@@ -34,14 +34,15 @@
   "dependencies": {
     "@pnpm/constants": "workspace:5.0.0",
     "@pnpm/error": "workspace:2.0.0",
-    "@pnpm/global-bin-dir": "workspace:3.0.0",
     "@pnpm/pnpmfile": "workspace:1.2.4",
     "@pnpm/read-project-manifest": "workspace:2.0.11",
     "@pnpm/types": "workspace:7.9.0",
     "@zkochan/npm-conf": "2.0.2",
     "camelcase": "^6.2.0",
+    "can-write-to-dir": "^1.1.1",
     "is-subdir": "^1.1.1",
     "normalize-registry-url": "2.0.0",
+    "path-name": "^1.0.0",
     "ramda": "^0.27.1",
     "realpath-missing": "^1.1.0",
     "which": "^2.0.2"
