--- packages/global-bin-dir/README.md
@@ -1,15 +1,0 @@
-# @pnpm/global-bin-dir
-
-> Finds a directory that is in PATH and we have permission to write to it
-
-[![npm version](https://img.shields.io/npm/v/@pnpm/global-bin-dir.svg)](https://www.npmjs.com/package/@pnpm/global-bin-dir)
-
-## Installation
-
-```sh
-pnpm add @pnpm/global-bin-dir
-```
-
-## License
-
-MIT
