--- packages/global-bin-dir/jest.config.js
@@ -1,1 +1,0 @@
-module.exports = require('../../jest.config.js')
