--- packages/global-bin-dir/package.json
@@ -1,41 +1,0 @@
-{
-  "name": "@pnpm/global-bin-dir",
-  "version": "3.0.0",
-  "description": "Finds a directory that is in PATH and we have permission to write to i",
-  "main": "lib/index.js",
-  "types": "lib/index.d.ts",
-  "files": [
-    "lib",
-    "!*.map"
-  ],
-  "engines": {
-    "node": ">=12.17"
-  },
-  "scripts": {
-    "lint": "eslint src/**/*.ts test/**/*.ts",
-    "test": "pnpm run compile && pnpm run _test",
-    "prepublishOnly": "pnpm run compile",
-    "compile": "tsc --build && pnpm run lint -- --fix",
-    "_test": "jest"
-  },
-  "repository": "https://github.com/pnpm/pnpm/blob/main/packages/global-bin-dir",
-  "keywords": [
-    "pnpm6",
-    "pnpm"
-  ],
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/pnpm/pnpm/issues"
-  },
-  "homepage": "https://github.com/pnpm/pnpm/blob/main/packages/global-bin-dir#readme",
-  "funding": "https://opencollective.com/pnpm",
-  "dependencies": {
-    "@pnpm/error": "workspace:2.0.0",
-    "can-write-to-dir": "^1.1.1",
-    "path-name": "^1.0.0"
-  },
-  "devDependencies": {
-    "@pnpm/global-bin-dir": "workspace:3.0.0",
-    "is-windows": "^1.0.2"
-  }
-}
