--- packages/global-bin-dir/tsconfig.json
@@ -1,16 +1,0 @@
-{
-  "extends": "@pnpm/tsconfig",
-  "compilerOptions": {
-    "outDir": "lib",
-    "rootDir": "src"
-  },
-  "include": [
-    "src/**/*.ts",
-    "../../typings/**/*.d.ts"
-  ],
-  "references": [
-    {
-      "path": "../error"
-    }
-  ]
-}
