--- packages/global-bin-dir/tsconfig.lint.json
@@ -1,8 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "include": [
-    "src/**/*.ts",
-    "test/**/*.ts",
-    "../../typings/**/*.d.ts"
-  ]
-}
